Updated base on feedback from Profile Setting Skeleton and EditProfile
