+++ client/src/App.tsx
@@ -5,8 +5,7 @@ import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom';
 import Login from './pages/Login/Login';
 import Signup from './pages/SignUp/SignUp';
 import Dashboard from './pages/Dashboard/Dashboard';
-import EditProfile from './pages/EditProfile/EditProfile';
-import ProfilePhoto from './pages/ProfilePhoto/ProfilePhoto';
+import Settings from './pages/Settings/Settings';
 import { AuthProvider } from './context/useAuthContext';
 import { SocketProvider } from './context/useSocketContext';
 import { SnackBarProvider } from './context/useSnackbarContext';
@@ -28,12 +27,7 @@ function App(): JSX.Element {
                 <Route exact path="/dashboard">
                   <Dashboard />
                 </Route>
-                <Route exact path={'/edit-profile'}>
-                  <EditProfile />
-                </Route>
-                <Route exact path={'/profile-photo'}>
-                  <ProfilePhoto />
-                </Route>
+                <Route path={'/settings'} component={Settings} />
                 <Route path="*">
                   <Redirect to="/login" />
                 </Route>
