--- client/src/components/BirthDate/useStyles.ts
@@ -1,5 +1,0 @@
-import { makeStyles } from '@mui/styles';
-
-const useStyles = makeStyles(() => ({}));
-
-export default useStyles;
